Django/requirements.txt
art==5.1
asgiref==3.3.4
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cycler==0.10.0
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
iso8601==0.1.13
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
m3u8==0.3.12
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.20.2
Pillow==8.1.0
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.0
scipy==1.6.1
six==1.15.0
sqlparse==0.4.1
twitch-dl==1.13.1
uritemplate==3.0.1
urllib3==1.26.2
